rshift.c: Replace assert with DEBUG and abort
assert in stdlib/rshift.c should be for debug purpose only and there is
no such check in any rshift assembly implementations nor in lshift.c.
This patch replaces assert with DEBUG and abort, similar to lshift.c
so that generic GMP codes from libc.a can be linked with libc.so in
atest-exp, atest-exp2 and atest-sincos, which depend on the GMP
implementation in glibc.

	* stdlib/rshift.c (mpn_rshift): Replace ssert with DEBUG and
	abort.